requirements.txt
pyyaml
awscli
openai
boto3
streamlit
pandas
